Add puppeteer for testing headless
Note:
Our task is to execute a JavaScript code on all frames, this must
also work for headless mode, so this puppeteer script is used to
test whether our patch satisfies the requirements or not.

We have been able to do that via RegisterExenstion(), but it only
works on headful mode. So we can't do that...

I need more reasearch on this!

Wish me good mental health :)

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>